better organise util functions
