+++ __tests__/lib/util/request.test.ts
@@ -1,41 +1,9 @@
 import {
   base64EncodeString,
   fetchJsonData,
-  formatDate,
-  formatDateHtml,
-  formatDateHtmlReverse,
   isJpegOrPngDataURL,
-  multipleClasses,
   postJsonData,
-} from "@/lib/util"
-
-describe("formatDate", () => {
-  it("formats a date as dd/mm/yyyy", () => {
-    expect(formatDate(new Date("2023/04/20"))).toEqual("20/04/2023")
-    expect(formatDate(new Date("1995/10/05"))).toEqual("05/10/1995")
-  })
-})
-
-describe("formatDateHtml", () => {
-  it("formats a date as yyyy-mm-dd", () => {
-    expect(formatDateHtml(new Date("2023/04/20"))).toEqual("20-04-2023")
-    expect(formatDateHtml(new Date("1995/10/05"))).toEqual("05-10-1995")
-  })
-})
-
-describe("formatDateHtmlReverse", () => {
-  it("formats a date as yyyy-mm-dd", () => {
-    expect(formatDateHtmlReverse(new Date("2023/04/20"))).toEqual("2023-04-20")
-    expect(formatDateHtmlReverse(new Date("1995/10/05"))).toEqual("1995-10-05")
-  })
-})
-
-describe("multipleClasses", () => {
-  it("combines multiple class names into a single string", () => {
-    expect(multipleClasses("foo", "bar", undefined, "", "baz")).toEqual("foo bar baz")
-    expect(multipleClasses(undefined, "bar", "", "baz")).toEqual("bar baz")
-  })
-})
+} from "@/lib/util/request"
 
 describe("fetchJsonData", () => {
   let mockFetch: jest.MockedFunction<typeof fetch>
