+++ src/components/receipt.tsx
@@ -12,7 +12,8 @@ import {
   Image as PdfImage,
 } from "@react-pdf/renderer"
 
-import { formatDate, multipleClasses } from "@/lib/util"
+import { multipleClasses } from "@/lib/util/etc"
+import { formatDate } from "@/lib/util/date"
 import { CompanyInfo, Donation } from "@/lib/qbo-api"
 
 export function HtmlReceipt({
